implement medium compliance
