+++ cmd/import/import.go
@@ -26,10 +26,8 @@
 package imprt
 
 import (
-	"bufio"
 	"fmt"
 	"log"
-	"os"
 	"strings"
 
 	"github.com/Cray-HPE/cani/internal/config"
@@ -37,7 +35,6 @@ import (
 	"github.com/Cray-HPE/cani/pkg/datastores"
 	"github.com/Cray-HPE/cani/pkg/devicetypes"
 	"github.com/Cray-HPE/cani/pkg/visual"
-	"github.com/google/uuid"
 	"github.com/spf13/cobra"
 )
 
@@ -233,469 +230,5 @@ func initializeDatastore(ctx *etlContext) error {
 		visual.PrintWarning("No existing devices found - prior phases may not have been run", ctx.opts)
 	}
 
-	return nil
-}
-
-// runExtractPhase executes the Extract phase of the ETL pipeline.
-// It calls provider.Import() which populates ctx.inventory with raw data from the source.
-func runExtractPhase(ctx *etlContext) error {
-	if ctx.debug {
-		visual.PrintPhaseHeader("EXTRACT", ctx.opts)
-		visual.PrintCaniOperation("Starting extract phase", ctx.opts)
-	}
-
-	if stepFlag {
-		if err := promptPhaseStart("EXTRACT", ctx.opts); err != nil {
-			return fmt.Errorf("extract phase: %w", err)
-		}
-	}
-
-	importer, ok := ctx.provider.(provider.Importer)
-	if !ok {
-		if ctx.debug {
-			visual.PrintPhaseComplete("EXTRACT", ctx.opts)
-		}
-		return nil
-	}
-
-	if ctx.debug {
-		visual.PrintProviderOperation(fmt.Sprintf("Importing data from provider %s", ctx.provider.Slug()), ctx.opts)
-	}
-
-	if err := importer.Import(ctx.cmd, ctx.args, ctx.inventory); err != nil {
-		return fmt.Errorf("extract phase: failed to import from %s: %w", ctx.provider.Slug(), err)
-	}
-
-	if ctx.debug {
-		visual.PrintPhaseComplete("EXTRACT", ctx.opts)
-	}
-	return nil
-}
-
-// runTransformPhase executes the Transform phase of the ETL pipeline.
-// It calls provider.Transform() and merges the result into ctx.inventory.
-func runTransformPhase(ctx *etlContext) error {
-	if ctx.debug {
-		visual.PrintPhaseHeader("TRANSFORM", ctx.opts)
-		visual.PrintCaniOperation("Starting transform phase", ctx.opts)
-	}
-
-	if stepFlag {
-		if err := promptPhaseStart("TRANSFORM", ctx.opts); err != nil {
-			return fmt.Errorf("transform phase: %w", err)
-		}
-	}
-
-	if ctx.debug {
-		visual.PrintProviderOperation(fmt.Sprintf("Transforming data from provider %s", ctx.provider.Slug()), ctx.opts)
-	}
-
-	result, err := ctx.provider.Transform(*ctx.inventory)
-	if err != nil {
-		return fmt.Errorf("transform phase: failed to transform data from %s: %w", ctx.provider.Slug(), err)
-	}
-
-	displayTransformSummary(ctx, result)
-	mergeTransformResult(ctx, result)
-
-	if ctx.debug {
-		visual.PrintPhaseComplete("TRANSFORM", ctx.opts)
-	}
-	return nil
-}
-
-// displayTransformSummary shows the transform results based on current mode.
-func displayTransformSummary(ctx *etlContext, result *devicetypes.TransformResult) {
-	if stepFlag {
-		displayBriefSummary(result, ctx.opts)
-	} else if ctx.debug {
-		summary := buildImportSummary(result.Racks, result.Devices, result.Cables)
-		visual.PrintImportSummary(summary, ctx.opts, false)
-	}
-}
-
-// mergeTransformResult merges all transformed entities into ctx.inventory.
-func mergeTransformResult(ctx *etlContext, result *devicetypes.TransformResult) {
-	result.EnsureUniqueDeviceNames()
-	mergeMetadata(ctx, result.Metadata)
-	locationRemap := mergeLocations(ctx, result.Locations)
-	rackRemap := mergeRacks(ctx, result.Racks)
-	remapDeviceParents(result.Devices, locationRemap, rackRemap)
-	printImportDiff(ctx, result.Devices)
-	mergeDevices(ctx, result.Devices)
-	mergeModules(ctx, result.Modules)
-	mergeCables(ctx, result.Cables)
-	mergeFrus(ctx, result.Frus)
-
-	// Single verify pass after all merges — avoids duplicate warnings
-	// from per-entity verify calls.
-	log.Printf("Verifying parent-child relationships")
-	ctx.inventory.VerifyParentChildRelationships()
-}
-
-// mergeMetadata adds roles, statuses, and tags from the transform result
-// into the inventory metadata catalog. Duplicates are silently ignored.
-func mergeMetadata(ctx *etlContext, meta *devicetypes.InventoryMetadata) {
-	if meta == nil {
-		return
-	}
-	for _, role := range meta.Roles {
-		_ = ctx.inventory.AddMetadata("roles", role)
-	}
-	for _, status := range meta.Statuses {
-		_ = ctx.inventory.AddMetadata("statuses", status)
-	}
-	for _, tag := range meta.Tags {
-		_ = ctx.inventory.AddMetadata("tags", tag)
-	}
-}
-
-// remapDeviceParents rewrites device Parent fields using the UUID remap
-// maps returned by MergeLocations and MergeRacks. This ensures devices
-// point to existing inventory UUIDs rather than ephemeral transform UUIDs.
-func remapDeviceParents(
-	devices map[uuid.UUID]*devicetypes.CaniDeviceType,
-	locationRemap, rackRemap map[uuid.UUID]uuid.UUID,
-) {
-	for _, dev := range devices {
-		if dev == nil || dev.Parent == uuid.Nil {
-			continue
-		}
-		if mapped, ok := rackRemap[dev.Parent]; ok {
-			dev.Parent = mapped
-		} else if mapped, ok := locationRemap[dev.Parent]; ok {
-			dev.Parent = mapped
-		}
-	}
-}
-
-// printImportDiff prints a summary of fields that will change when incoming
-// devices are merged into the existing inventory.
-func printImportDiff(ctx *etlContext, incoming map[uuid.UUID]*devicetypes.CaniDeviceType) {
-	type fieldDiff struct {
-		field  string
-		oldVal string
-		newVal string
-	}
-
-	var changedDevices []struct {
-		name  string
-		diffs []fieldDiff
-	}
-
-	for _, dev := range incoming {
-		if dev == nil || dev.Name == "" {
-			continue
-		}
-		// Find the existing device by name.
-		var existing *devicetypes.CaniDeviceType
-		for _, e := range ctx.inventory.Devices {
-			if e != nil && e.Name == dev.Name {
-				existing = e
-				break
-			}
-		}
-		if existing == nil {
-			continue
-		}
-
-		var diffs []fieldDiff
-		if dev.RackPosition != 0 && dev.RackPosition != existing.RackPosition {
-			diffs = append(diffs, fieldDiff{"position", fmt.Sprintf("%d", existing.RackPosition), fmt.Sprintf("%d", dev.RackPosition)})
-		}
-		if dev.Status != "" && dev.Status != existing.Status {
-			diffs = append(diffs, fieldDiff{"status", existing.Status, dev.Status})
-		}
-		if dev.Role != "" && dev.Role != existing.Role {
-			diffs = append(diffs, fieldDiff{"role", existing.Role, dev.Role})
-		}
-		if dev.Model != "" && dev.Model != existing.Model {
-			diffs = append(diffs, fieldDiff{"model", existing.Model, dev.Model})
-		}
-		if dev.Parent != uuid.Nil && dev.Parent != existing.Parent {
-			diffs = append(diffs, fieldDiff{"parent", existing.Parent.String(), dev.Parent.String()})
-		}
-		if dev.Face != "" && dev.Face != existing.Face {
-			diffs = append(diffs, fieldDiff{"face", existing.Face, dev.Face})
-		}
-		if dev.Serial != "" && dev.Serial != existing.Serial {
-			diffs = append(diffs, fieldDiff{"serial", existing.Serial, dev.Serial})
-		}
-
-		if len(diffs) > 0 {
-			changedDevices = append(changedDevices, struct {
-				name  string
-				diffs []fieldDiff
-			}{dev.Name, diffs})
-		}
-	}
-
-	if len(changedDevices) == 0 {
-		return
-	}
-
-	log.Println()
-	log.Println("=== Devices with pending changes ===")
-	for _, cd := range changedDevices {
-		log.Printf("  %s: %d field(s) changed:", cd.name, len(cd.diffs))
-		for _, d := range cd.diffs {
-			log.Printf("      %s: %s --> %s", d.field, d.oldVal, d.newVal)
-		}
-	}
-	log.Println()
-}
-
-// mergeLocations adds transformed locations to the inventory.
-func mergeLocations(ctx *etlContext, locations map[uuid.UUID]*devicetypes.CaniLocationType) map[uuid.UUID]uuid.UUID {
-	if len(locations) == 0 {
-		return nil
-	}
-	if ctx.debug {
-		visual.PrintCaniOperation(fmt.Sprintf("Merging %d locations into inventory", len(locations)), ctx.opts)
-	}
-	return ctx.inventory.MergeLocations(locations)
-}
-
-// mergeRacks adds transformed racks to the inventory.
-func mergeRacks(ctx *etlContext, racks map[uuid.UUID]*devicetypes.CaniRackType) map[uuid.UUID]uuid.UUID {
-	if len(racks) == 0 {
-		return nil
-	}
-	if ctx.debug {
-		visual.PrintCaniOperation(fmt.Sprintf("Merging %d racks into inventory", len(racks)), ctx.opts)
-	}
-	return ctx.inventory.MergeRacks(racks)
-}
-
-// mergeDevices adds transformed devices to the inventory.
-// In strict mode, unclassified devices (no slug/model) are rejected.
-// In step mode, the user is prompted to interactively classify them.
-func mergeDevices(ctx *etlContext, devices map[uuid.UUID]*devicetypes.CaniDeviceType) {
-	if len(devices) == 0 {
-		return
-	}
-	if ctx.debug {
-		visual.PrintCaniOperation(fmt.Sprintf("Merging %d transformed devices into inventory", len(devices)), ctx.opts)
-	}
-
-	strict := config.Cfg.Strict
-	skipped := ctx.inventory.MergeDevicesStrict(devices, strict)
-
-	if len(skipped) == 0 {
-		return
-	}
-
-	// In step mode, prompt user to classify each skipped device.
-	if stepFlag {
-		classifyOpts := devicetypes.ClassifyOptions{NoColor: noColorFlag}
-		classified := 0
-		for _, ud := range skipped {
-			slug, err := devicetypes.PromptForDeviceType(ud, classifyOpts)
-			if err != nil {
-				log.Printf("  ! %s: classification error: %v", ud.Name, err)
-				continue
-			}
-			device := devices[ud.ID]
-			if device == nil {
-				continue
-			}
-			if slug == "" {
-				log.Printf("  - %s: skipped (no type selected)", ud.Name)
-			} else {
-				if err := devicetypes.ApplyDeviceType(device, slug); err != nil {
-					log.Printf("  ! %s: failed to apply type %q: %v", ud.Name, slug, err)
-				} else {
-					classified++
-				}
-			}
-			// Always merge the device so modules/FRUs can reference it as a parent.
-			ctx.inventory.MergeDevices(map[uuid.UUID]*devicetypes.CaniDeviceType{ud.ID: device})
-		}
-		if classified > 0 {
-			log.Printf("  Classified %d of %d unclassified devices", classified, len(skipped))
-		}
-		return
-	}
-
-	// Non-interactive: warn and suggest the classify command.
-	log.Printf("")
-	log.Printf("  ⚠ %d devices rejected (no device type slug or model):", len(skipped))
-	for _, ud := range skipped {
-		log.Printf("    - %s", ud.Name)
-	}
-	log.Printf("")
-	log.Printf("  To assign types interactively, run:")
-	log.Printf("    cani alpha classify")
-	log.Printf("  Or re-import with --step to classify inline.")
-	log.Printf("  To allow unclassified devices, use --strict=false")
-	log.Printf("")
-
-	// Always merge skipped devices so modules/FRUs can reference them as parents.
-	for _, ud := range skipped {
-		device := devices[ud.ID]
-		if device == nil {
-			continue
-		}
-		ctx.inventory.MergeDevices(map[uuid.UUID]*devicetypes.CaniDeviceType{ud.ID: device})
-	}
-}
-
-// mergeModules adds transformed modules to the inventory.
-func mergeModules(ctx *etlContext, modules map[uuid.UUID]*devicetypes.CaniModuleType) {
-	if len(modules) == 0 {
-		return
-	}
-	if ctx.debug {
-		visual.PrintCaniOperation(fmt.Sprintf("Merging %d modules into inventory", len(modules)), ctx.opts)
-	}
-	ctx.inventory.MergeModules(modules)
-}
-
-// mergeCables adds transformed cables to the inventory.
-func mergeCables(ctx *etlContext, cables map[uuid.UUID]*devicetypes.CaniCableType) {
-	if len(cables) == 0 {
-		return
-	}
-	if ctx.debug {
-		visual.PrintCaniOperation(fmt.Sprintf("Merging %d cables into inventory", len(cables)), ctx.opts)
-	}
-	ctx.inventory.MergeCables(cables)
-}
-
-// mergeFrus adds transformed FRUs to the inventory.
-func mergeFrus(ctx *etlContext, frus map[uuid.UUID]*devicetypes.CaniFruType) {
-	if len(frus) == 0 {
-		return
-	}
-	if ctx.debug {
-		visual.PrintCaniOperation(fmt.Sprintf("Merging %d FRUs into inventory", len(frus)), ctx.opts)
-	}
-	ctx.inventory.MergeFrus(frus)
-}
-
-// runLoadPhase executes the Load phase of the ETL pipeline.
-// It persists ctx.inventory to the local datastore.
-func runLoadPhase(ctx *etlContext) error {
-	if ctx.debug {
-		visual.PrintPhaseHeader("LOAD", ctx.opts)
-		visual.PrintCaniOperation("Starting load phase", ctx.opts)
-	}
-
-	if stepFlag {
-		if err := promptPhaseStart("LOAD", ctx.opts); err != nil {
-			return fmt.Errorf("load phase: %w", err)
-		}
-	}
-
-	if ctx.debug {
-		visual.PrintCaniOperation("Saving inventory to local datastore", ctx.opts)
-	}
-
-	if err := datastores.Datastore.Save(ctx.inventory); err != nil {
-		return fmt.Errorf("load phase: failed to save inventory: %w", err)
-	}
-
-	if ctx.debug {
-		visual.PrintPhaseComplete("LOAD", ctx.opts)
-	}
-	return nil
-}
-
-// buildImportSummary creates a summary from transformed racks, devices, and cables.
-func buildImportSummary(
-	racks map[uuid.UUID]*devicetypes.CaniRackType,
-	devices map[uuid.UUID]*devicetypes.CaniDeviceType,
-	cables map[uuid.UUID]*devicetypes.CaniCableType,
-) visual.ImportSummary {
-	summary := visual.ImportSummary{
-		RackNames:     []string{},
-		DevicesByRack: make(map[string][]string),
-		Cables:        []visual.CableSummary{},
-	}
-
-	// Build rack ID to name lookup
-	racksByID := make(map[uuid.UUID]string)
-	for id, rack := range racks {
-		summary.RackNames = append(summary.RackNames, rack.Name)
-		racksByID[id] = rack.Name
-	}
-
-	// Group devices by their ancestor rack (walk parent chain)
-	for _, device := range devices {
-		rackName := findAncestorRack(device, devices, racksByID)
-		summary.DevicesByRack[rackName] = append(summary.DevicesByRack[rackName], device.Name)
-	}
-
-	// Add cable count
-	for _, cable := range cables {
-		summary.Cables = append(summary.Cables, visual.CableSummary{
-			CableType: cable.Slug,
-		})
-	}
-
-	return summary
-}
-
-// findAncestorRack walks the parent chain of a device to find its ancestor rack.
-// Returns the rack name if found, or empty string if the device is standalone.
-func findAncestorRack(
-	device *devicetypes.CaniDeviceType,
-	devices map[uuid.UUID]*devicetypes.CaniDeviceType,
-	racksByID map[uuid.UUID]string,
-) string {
-	visited := make(map[uuid.UUID]bool)
-	current := device.Parent
-	for current != uuid.Nil && !visited[current] {
-		visited[current] = true
-		if name, ok := racksByID[current]; ok {
-			return name
-		}
-		parent, ok := devices[current]
-		if !ok {
-			break
-		}
-		current = parent.Parent
-	}
-	return ""
-}
-
-// promptPhaseStart displays a visual checkpoint and waits for Enter to begin a phase
-func promptPhaseStart(phase string, opts visual.ETLOptions) error {
-	cyan, yellow, gray := visual.ColorCyan, visual.ColorYellow, visual.ColorGray
-	reset := visual.ColorReset
-	if opts.NoColor {
-		cyan, yellow, gray, reset = "", "", "", ""
-	}
-
-	// Calculate padding for phase name to align box
-	phaseText := fmt.Sprintf("▶ Ready to begin %s phase", phase)
-	boxWidth := 50
-	padding := boxWidth - len(phaseText) - 2 // -2 for leading spaces
-	if padding < 0 {
-		padding = 0
-	}
-
-	fmt.Println()
-	fmt.Printf("%s┌%s┐%s\n", cyan, strings.Repeat("─", boxWidth), reset)
-	fmt.Printf("%s│%s  %s%s%s%s%s│%s\n", cyan, reset, yellow, phaseText, reset, strings.Repeat(" ", padding), cyan, reset)
-	fmt.Printf("%s│%s%s%s│%s\n", cyan, reset, strings.Repeat(" ", boxWidth), cyan, reset)
-	fmt.Printf("%s│%s  %sPress Enter to continue, or Ctrl+C to abort...%s  %s│%s\n", cyan, reset, gray, reset, cyan, reset)
-	fmt.Printf("%s└%s┘%s\n", cyan, strings.Repeat("─", boxWidth), reset)
-
-	reader := bufio.NewReader(os.Stdin)
-	_, err := reader.ReadString('\n')
-	if err != nil {
-		return fmt.Errorf("failed to read input: %w", err)
-	}
-	return nil
-}
-
-// displayBriefSummary shows a summary of transformed items
-func displayBriefSummary(result *devicetypes.TransformResult, opts visual.ETLOptions) error {
-	summary := buildImportSummary(result.Racks, result.Devices, result.Cables)
-	visual.PrintImportSummary(summary, opts, true)
-
-	fmt.Printf("\nTransformed: %d racks, %d devices, %d cables\n",
-		len(result.Racks), len(result.Devices), len(result.Cables))
 	return nil
 }
