+++ cmd/show/nouns.go
@@ -201,146 +201,3 @@ func showRacks(cmd *cobra.Command, args []string) error {
 		return marshalAndPrint(racks)
 	}
 }
-
-// newDeviceShowCommand creates the "show device" subcommand.
-func newDeviceShowCommand() *cobra.Command {
-	return &cobra.Command{
-		Use:   "device [name|uuid]",
-		Short: "List devices in the inventory.",
-		Long:  "List devices, or show a single device by name or UUID.",
-		Args:  cobra.MaximumNArgs(1),
-		RunE:  showDevices,
-	}
-}
-
-func showDevices(cmd *cobra.Command, args []string) error {
-	inv, err := loadInventory(cmd, args)
-	if err != nil {
-		return err
-	}
-
-	if len(args) == 1 {
-		return showSingleDevice(cmd, inv, args[0])
-	}
-
-	devices := make([]*devicetypes.CaniDeviceType, 0, len(inv.Devices))
-	for _, device := range inv.Devices {
-		devices = append(devices, device)
-	}
-
-	sortKey, _ := cmd.Flags().GetString("sort")
-	sort.Slice(devices, func(i, j int) bool {
-		switch sortKey {
-		case "type":
-			return devices[i].Type < devices[j].Type
-		case "id":
-			return devices[i].ID.String() < devices[j].ID.String()
-		case "status":
-			return devices[i].Status < devices[j].Status
-		case "vendor":
-			return devices[i].GetVendor() < devices[j].GetVendor()
-		case "model":
-			return devices[i].Model < devices[j].Model
-		default:
-			return devices[i].Name < devices[j].Name
-		}
-	})
-
-	format, _ := cmd.Flags().GetString("format")
-	switch format {
-	case "table":
-		visual.PrintDeviceTable(devices, inv, visual.TreeFilter{})
-		return nil
-	case "tree":
-		tf := treeFilterFromCmd(cmd)
-		nodes := visual.BuildDeviceTree(devices, inv, tf)
-		visual.RenderTreeOutput(nodes, tf.NoColor)
-		return nil
-	default:
-		return marshalAndPrint(devices)
-	}
-}
-
-func showSingleDevice(cmd *cobra.Command, inv *devicetypes.Inventory, arg string) error {
-	dev, err := findDeviceByNameOrUUID(arg, inv)
-	if err != nil {
-		return err
-	}
-	format, _ := cmd.Flags().GetString("format")
-	switch format {
-	case "table":
-		visual.PrintDeviceTable([]*devicetypes.CaniDeviceType{dev}, inv, visual.TreeFilter{})
-		return nil
-	case "tree":
-		tf := treeFilterFromCmd(cmd)
-		nodes := []visual.TreeNode{visual.DeviceToTreeNode(dev, inv, tf)}
-		visual.RenderTreeOutput(nodes, tf.NoColor)
-		return nil
-	default:
-		return marshalAndPrint(dev)
-	}
-}
-
-// newModuleCommand creates the "show module" subcommand.
-func newModuleCommand() *cobra.Command {
-	return &cobra.Command{
-		Use:   "module [name|uuid]",
-		Short: "List modules in the inventory.",
-		Long:  "List modules, or show a single module by name or UUID.",
-		Args:  cobra.MaximumNArgs(1),
-		RunE:  showModules,
-	}
-}
-
-func showModules(cmd *cobra.Command, args []string) error {
-	inv, err := loadInventory(cmd, args)
-	if err != nil {
-		return err
-	}
-
-	if len(args) == 1 {
-		return showSingleModule(cmd, inv, args[0])
-	}
-
-	modules := make([]*devicetypes.CaniModuleType, 0, len(inv.Modules))
-	for _, mod := range inv.Modules {
-		modules = append(modules, mod)
-	}
-	sort.Slice(modules, func(i, j int) bool {
-		return modules[i].Name < modules[j].Name
-	})
-
-	format, _ := cmd.Flags().GetString("format")
-	switch format {
-	case "table":
-		visual.PrintModuleTable(modules, inv)
-		return nil
-	case "tree":
-		tf := treeFilterFromCmd(cmd)
-		nodes := visual.BuildModuleTree(modules, inv, tf)
-		visual.RenderTreeOutput(nodes, tf.NoColor)
-		return nil
-	default:
-		return marshalAndPrint(modules)
-	}
-}
-
-func showSingleModule(cmd *cobra.Command, inv *devicetypes.Inventory, arg string) error {
-	mod, err := findModuleByNameOrUUID(arg, inv)
-	if err != nil {
-		return err
-	}
-	format, _ := cmd.Flags().GetString("format")
-	switch format {
-	case "table":
-		visual.PrintModuleTable([]*devicetypes.CaniModuleType{mod}, inv)
-		return nil
-	case "tree":
-		tf := treeFilterFromCmd(cmd)
-		nodes := []visual.TreeNode{visual.ModuleToTreeNode(mod, inv, tf)}
-		visual.RenderTreeOutput(nodes, tf.NoColor)
-		return nil
-	default:
-		return marshalAndPrint(mod)
-	}
-}
